% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteTFSRepository.R
\name{deleteTFSRepository}
\alias{deleteTFSRepository}
\title{[DEPRECATED] Delete a git repository in an existing TFS (Git) Project}
\usage{
deleteTFSRepository(URL, Credentials, Repo, ..., URLSub = "DefaultCollection")
}
\arguments{
\item{URL}{URL for root TFS e.g. http://stephlocke.visualstudio.com}

\item{Credentials}{\code{TFSAuth} object}

\item{Repo}{Name of repository to be deleted}

\item{...}{Additional arguments to httr::DELETE}

\item{URLSub}{Modify this if you need to reference somewhere other than DefaultCollection}
}
\value{
Result    String containing delete status
}
\description{
This attempts to delete a repository and returns whether the operation was successful
}
\keyword{list}
\keyword{repository}
\keyword{tfs}

