% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTFSProjects.R
\name{getTFSProjects}
\alias{getTFSProjects}
\title{Get existing TFS (Git) Projects}
\usage{
getTFSProjects(URL, Credentials, ..., Flatten = TRUE,
  URLSub = "DefaultCollection")
}
\arguments{
\item{URL}{URL for root TFS e.g. http://stephlocke.visualstudio.com}

\item{Credentials}{\code{TFSAuth} object}

\item{...}{Additional arguments to httr::GET}

\item{Flatten}{If Flatten=FALSE, the results will be returned in the native list format}

\item{URLSub}{Modify this if you need to reference somewhere other than DefaultCollection}
}
\value{
projects    Data.table (unless Flatten=FALSE, in which case a list) with each project, GUID, & type
}
\description{
Find out what Projects & repositories exist on TFS.
This will help you isolate the GUIDs etc needed to add 
new Git repositories.
}
\examples{
tfs           <- "https://stephlocke.visualstudio.com"
authcreds     <- TFSAuth(pwd="fz43enydh7vi2o6jqir2gmftohh7ooz2lizqvy6jxtw4ltrpwola")
tfsprojects   <- getTFSProjects(tfs,authcreds) 

}
\seealso{
\code{\link{createTFSRepository}}
}
\keyword{list}
\keyword{repository}
\keyword{tfs}

