% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFSAuth.R
\name{TFSAuth}
\alias{TFSAuth}
\title{Use alternate authentication credentials or personal access
tokens to connect to your TFS instance.}
\usage{
TFSAuth(user = "", pwd)
}
\arguments{
\item{user}{Username (no need to populate if using public access token)}

\item{pwd}{Password or public access token}
}
\value{
auth headers
}
\description{
Use alternate authentication credentials or personal access
tokens to connect to your TFS instance.
}
\examples{
authcreds <- TFSAuth(pwd="fz43enydh7vi2o6jqir2gmftohh7ooz2lizqvy6jxtw4ltrpwola")
}

