% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfsR-package.r
\docType{package}
\name{tfsR}
\alias{tfsR}
\alias{tfsR-package}
\title{tfsR}
\description{
{
[![Build Status](https://travis-ci.org/stephlocke/tfsR.png?branch=dev)](https://travis-ci.org/stephlocke/tfsR)[![Coverage Status](https://coveralls.io/repos/stephlocke/tfsR/badge.svg)](https://coveralls.io/r/stephlocke/tfsR)

This package allows you to work with Visual Studio Team Services (VSTS) and Team Foundation Server (TFS). 
The original version(s) of the package were solely aimed creating git repositories but with the ability 
to create Projects and more, [@velincw](https://github.com/vergilcw) to be a little bit more ambitious! 
This will be a complete rebuild, with some aliases provided for backwards compatability.
}
}

