% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createTFSRepository.R
\name{createTFSRepository}
\alias{createTFSRepository}
\title{Setup a git repository in an existing TFS (Git) Project}
\usage{
createTFSRepository(URL, Credentials, ParentProject, NewRepo, ...,
  URLSub = "DefaultCollection")
}
\arguments{
\item{URL}{URL for root TFS e.g. http://stefflocke.visualstudio.com}

\item{Credentials}{`httr::authenticate("username","password")` object}

\item{ParentProject}{Name of parent project to store repository in}

\item{NewRepo}{Name for new repository}

\item{...}{Additional arguments to httr::POST}

\item{URLSub}{Modify this if you need to reference somewhere other than DefaultCollection}
}
\value{
RepoURL    String containing location of repository
}
\description{
This produces a repository and returns the URL that can
be used for pushing/pulling
}
\examples{
# simple example
tfs          <- "https://stefflocke.visualstudio.com"
authcreds    <- httr::authenticate("tfsexample","UsedForExampl3s")
parentproj   <- "GitRepoContainer"
newrepo      <- as.character(random::randomStrings(n=1, len=6))

createdrepo  <- createTFSRepository(tfs,authcreds,parentproj,newrepo)
}
\keyword{list}
\keyword{repository}
\keyword{tfs}

