% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteTFSRepository.R
\name{deleteTFSRepository}
\alias{deleteTFSRepository}
\title{Delete a git repository in an existing TFS (Git) Project}
\usage{
deleteTFSRepository(URL, Credentials, Repo, ..., URLSub = "DefaultCollection")
}
\arguments{
\item{URL}{URL for root TFS e.g. http://stephlocke.visualstudio.com}

\item{Credentials}{\code{TFSAuth} object}

\item{Repo}{Name of repository to be deleted}

\item{...}{Additional arguments to httr::DELETE}

\item{URLSub}{Modify this if you need to reference somewhere other than DefaultCollection}
}
\value{
Result    String containing delete status
}
\description{
This attempts to delete a repository and returns whether the operation was successful
}
\examples{
tfs        <- "https://stephlocke.visualstudio.com"
authcreds  <- TFSAuth(pwd="fz43enydh7vi2o6jqir2gmftohh7ooz2lizqvy6jxtw4ltrpwola")
parentproj <- "tfsr3"
newrepo    <- as.character(random::randomStrings(n=1, len=6))
createdrepo  <- createTFSRepository(tfs,authcreds,parentproj,newrepo)

deleteTFSRepository(tfs,authcreds,newrepo)
}
\keyword{list}
\keyword{repository}
\keyword{tfs}

