% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/createTFSRepository.R, R/getTFSProjects.R
\name{getTFSProjects}
\alias{getTFSProjects}
\title{Get existing TFS (Git) Projects}
\usage{
getTFSProjects(URL, Credentials, ..., Flatten = TRUE,
  URLSub = "DefaultCollection")

getTFSProjects(URL, Credentials, ..., Flatten = TRUE,
  URLSub = "DefaultCollection")
}
\arguments{
\item{URL}{URL for root TFS e.g. http://stefflocke.visualstudio.com}

\item{Credentials}{`httr::authenticate("username","password")` object}

\item{...}{Additional arguments to httr::GET}

\item{Flatten}{If Flatten=FALSE, the results will be returned in the native list format}

\item{URLsub}{Modify this if you need to reference somewhere other than DefaultCollection}

\item{URL}{URL for root TFS e.g. http://stefflocke.visualstudio.com}

\item{Credentials}{`httr::authenticate("username","password")` object}

\item{...}{Additional arguments to httr::GET}

\item{Flatten}{If Flatten=FALSE, the results will be returned in the native list format}

\item{URLsub}{Modify this if you need to reference somewhere other than DefaultCollection}
}
\value{
projects    Data.table (unless Flatten=FALSE, in which case a list) with each project, GUID, & type

projects    Data.table (unless Flatten=FALSE, in which case a list) with each project, GUID, & type
}
\description{
Find out what Projects & repositories exist on TFS.
This will help you isolate the GUIDs etc needed to add
new Git repositories.

Find out what Projects & repositories exist on TFS.
This will help you isolate the GUIDs etc needed to add
new Git repositories.
}
\examples{
# simple example
tfs<-"https://stefflocke.visualstudio.com"
authcreds<-httr::authenticate("stephstemp@grr.la","UsedForExamples")
tfsprojects<-getTFSProjects(tfs,authcreds)
# simple example
tfs<-"https://stefflocke.visualstudio.com"
authcreds<-httr::authenticate("stephstemp@grr.la","UsedForExamples")
tfsprojects<-getTFSProjects(tfs,authcreds)
}
\seealso{
\code{\link{bar}}

\code{\link{bar}}
}
\keyword{bar}
\keyword{foo}

