% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFSAuth.R
\name{TFSAuth}
\alias{TFSAuth}
\title{[DEPRECATED] Use alternate authentication credentials or personal access
tokens to connect to your TFS instance.}
\usage{
TFSAuth(user = "", pwd, type = NULL)
}
\arguments{
\item{user}{Username (no need to populate if using public access token)}

\item{pwd}{Password or public access token}

\item{type}{Authentication type, passed into `httr::authenticate`}
}
\value{
auth headers
}
\description{
[DEPRECATED] Use alternate authentication credentials or personal access
tokens to connect to your TFS instance.
}

